refactor azure package in to smaller files
